Renamed burn and operatorBurn to redeem and operatorRedeem: L04

+++ contracts/compound/ICErc20.sol
@@ -20,8 +20,8 @@ pragma solidity 0.5.12;
 
 contract ICErc20 {
     address public underlying;
-    function mint(uint mintAmount) external returns (uint);
-    function redeemUnderlying(uint redeemAmount) external returns (uint);
+    function mint(uint256 mintAmount) external returns (uint);
+    function redeemUnderlying(uint256 redeemAmount) external returns (uint);
     function balanceOfUnderlying(address owner) external returns (uint);
     function getCash() external view returns (uint);
     function supplyRatePerBlock() external view returns (uint);
